# cmd/api/docs/swagger.yaml
basePath: /
definitions:
...
  clap_internal_modules_user_dto.UpdateMobileProfileRequest:
    properties:
      email:
        type: string
      name:
        maxLength: 100
        minLength: 1
        type: string
    type: object
  clap_internal_shared_response.ErrorDetail:
    properties:
      code:
        type: integer
      details:
        items:
          type: string
        type: array
      message:
        type: string
    type: object
  clap_internal_shared_response.Response:
    properties:
      data: {}
      error:
        $ref: '#/definitions/clap_internal_shared_response.ErrorDetail'
      message:
        type: string
      success:
        type: boolean
    type: object
...
  internal_modules_user_handler.CreateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      city:
        type: string
      country:
        type: string
      date_of_birth:
        type: string
    type: object
  internal_modules_user_handler.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      bio:
        type: string
      city:
        type: string
      country:
        type: string
      date_of_birth:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Clap API Support
  description: Clap match-day experience backend (auth, clubs, matches, realtime,
    shop).
  license:
    name: Proprietary
  termsOfService: http://swagger.io/terms/
  title: Clap Backend API
  version: "1.0"
paths:
...
  /api/v1/profile/leaderboard:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Leaderboard
      tags:
      - mobile-profile
  /api/v1/profile/me:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Mobile profile me
      tags:
      - mobile-profile
    patch:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/clap_internal_modules_user_dto.UpdateMobileProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update mobile profile
      tags:
      - mobile-profile
  /api/v1/profile/me/avatar:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Avatar image
        in: formData
        name: file
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Upload avatar
      tags:
      - mobile-profile
  /api/v1/profiles/me:
    delete:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Delete profile
      tags:
      - profiles
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Get profile
      tags:
      - profiles
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user_handler.CreateProfileRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Create profile
      tags:
      - profiles
    put:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/internal_modules_user_handler.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Update profile
      tags:
      - profiles
...
  /api/v1/videos:
    post:
      consumes:
      - multipart/form-data
      parameters:
      - description: Video file
        in: formData
        name: file
        required: true
        type: file
      - description: Caption
        in: formData
        name: caption
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Upload video
      tags:
      - videos
  /api/v1/videos/{video_id}/like:
    delete:
      parameters:
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Unlike video
      tags:
      - videos
    post:
      parameters:
      - description: Video ID
        in: path
        name: video_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Like video
      tags:
      - videos
  /api/v1/videos/feed:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: Video feed
      tags:
      - videos
  /api/v1/videos/mine:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/clap_internal_shared_response.Response'
      security:
      - BearerAuth: []
      summary: My videos
      tags:
      - videos
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'JWT access token. Example: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"